Switched type of stack index from usize to i64
This enables support for negative stack indices, which we will have
correspond to "stack + num", for accessing function arguments.

The type which is mapped to in `ctx` is also changed, so that we can use `ctx`
to store the location of function arguments.

+++ compiler/src/assembly.rs
@@ -13,7 +13,7 @@ pub enum Val {
 
     /// stack index, will be multiplied by 8 to find actual position since we index the stack in
     /// 64-bit words
-    StackIndex(usize),
+    StackIndex(i64),
 }
 
 impl Display for Val {
